Refactor utilities into separate editor files
