+++ Editor/SmartBackups.cs
@@ -8,9 +8,7 @@ using System.IO;
 using System.Linq;
 using System.Text;
 using System.Collections.Generic;
-using System.Security.Cryptography;
 using System.IO.Compression;
-using System.Reflection;
 using System.Diagnostics;
 using System.Threading;
 using System.Threading.Tasks;
@@ -20,171 +18,6 @@ using UnityEngine;
 
 namespace AvatarSmartBackup
 {
-    [InitializeOnLoad]
-    internal static class MainThread
-    {
-        static readonly int MainId;
-        static MainThread()
-        {
-            MainId = Thread.CurrentThread.ManagedThreadId;
-        }
-        public static void Invoke(Action a)
-        {
-            if (Thread.CurrentThread.ManagedThreadId == MainId) a();
-            else EditorApplication.delayCall += () => a();
-        }
-        public static T InvokeBlocking<T>(Func<T> f)
-        {
-            if (Thread.CurrentThread.ManagedThreadId == MainId) return f();
-            T result = default;
-            var ev = new ManualResetEventSlim();
-            EditorApplication.delayCall += () => { result = f(); ev.Set(); };
-            ev.Wait();
-            return result;
-        }
-    }
-
-    internal static class Log
-    {
-        const string Tag = "[ Avatar Backup System ] ";
-        static readonly string FilePath = Path.Combine(FileUtilEx.BackupRoot, "asb.log");
-        static void Write(string level, string msg)
-        {
-            string line = $"{DateTime.Now:yyyy-MM-dd HH:mm:ss} [{level}] {msg}";
-            try
-            {
-                Directory.CreateDirectory(Path.GetDirectoryName(FilePath));
-                File.AppendAllText(FilePath, line + Environment.NewLine);
-            }
-            catch { }
-        }
-        public static void Info(string msg)
-        {
-            UnityEngine.Debug.Log(Tag + msg);
-            Write("INFO", msg);
-        }
-        public static void Warn(string msg)
-        {
-            UnityEngine.Debug.LogWarning(Tag + msg);
-            Write("WARN", msg);
-        }
-        public static void Err(string msg)
-        {
-            UnityEngine.Debug.LogError(Tag + msg);
-            Write("ERROR", msg);
-        }
-    }
-
-    internal static class ProgressUX
-    {
-        // Wrapper sullo UnityEditor.Progress (non modale). Fallback a nessuna UI se assente.
-        static Type _progressType;
-        static MethodInfo _start, _report, _finish, _registerCancel, _remove;
-        static bool _checked;
-
-        static void Ensure()
-        {
-            if (_checked) return;
-            _checked = true;
-            _progressType = Type.GetType("UnityEditor.Progress, UnityEditor");
-            if (_progressType == null) return;
-            _start = _progressType.GetMethod("Start", new[] { typeof(string), typeof(string), typeof(UnityEditor.Progress.Options) });
-            if (_start == null) _start = _progressType.GetMethod("Start", new[] { typeof(string), typeof(string) });
-            _report = _progressType.GetMethod("Report", new[] { typeof(int), typeof(float), typeof(string) });
-            _finish = _progressType.GetMethod("Finish", new[] { typeof(int) });
-            // Try both signatures found across Unity versions: Func<bool> and Action
-            _registerCancel = _progressType.GetMethod("RegisterCancelCallback", new[] { typeof(int), typeof(Func<bool>) })
-                                ?? _progressType.GetMethod("RegisterCancelCallback", new[] { typeof(int), typeof(Action) });
-            _remove = _progressType.GetMethod("Remove", new[] { typeof(int) });
-        }
-
-        public static int Start(string title, string desc, bool cancellable, Func<bool> onCancel = null)
-        {
-            Ensure();
-            if (_progressType == null) return -1;
-            try
-            {
-                return MainThread.InvokeBlocking(() =>
-                {
-                    int id;
-                    if (_start != null && _start.GetParameters().Length == 3)
-                    {
-                        var opts = (UnityEditor.Progress.Options)Enum.Parse(typeof(UnityEditor.Progress.Options),
-                            cancellable ? "Managed" : "None");
-                        id = (int)_start.Invoke(null, new object[] { title, desc, opts });
-                    }
-                    else if (_start != null)
-                    {
-                        id = (int)_start.Invoke(null, new object[] { title, desc });
-                    }
-                    else { return -1; }
-                    if (onCancel != null && _registerCancel != null)
-                    {
-                        var pars = _registerCancel.GetParameters();
-                        if (pars.Length == 2 && pars[1].ParameterType == typeof(Action))
-                        {
-                            Action act = () => onCancel();
-                            _registerCancel.Invoke(null, new object[] { id, act });
-                        }
-                        else
-                        {
-                            _registerCancel.Invoke(null, new object[] { id, onCancel });
-                        }
-                    }
-                    return id;
-                });
-            }
-            catch { return -1; }
-        }
-
-        public static void Report(int id, float p, string desc)
-        {
-            if (id < 0) return;
-            Ensure();
-            try { MainThread.Invoke(() => _report?.Invoke(null, new object[] { id, Mathf.Clamp01(p), desc })); }
-            catch { /* ignore */ }
-        }
-
-        public static void Finish(int id)
-        {
-            if (id < 0) return;
-            Ensure();
-            try { MainThread.Invoke(() => _finish?.Invoke(null, new object[] { id })); }
-            catch { /* ignore */ }
-        }
-        public static void Remove(int id)
-        {
-            if (id < 0) return;
-            Ensure();
-            try { MainThread.Invoke(() => _remove?.Invoke(null, new object[] { id })); }
-            catch { /* ignore */ }
-        }
-    }
-
-    // Tracks editor activity (input, compilation, playmode) to detect idle periods
-    internal static class EditorIdle
-    {
-        static double _lastActivity = EditorApplication.timeSinceStartup;
-        static EditorIdle()
-        {
-            EditorApplication.update += Update;
-        }
-        static void Update()
-        {
-            if (UnityEngine.Input.anyKey ||
-                UnityEngine.Input.GetMouseButton(0) ||
-                UnityEngine.Input.GetMouseButton(1) ||
-                UnityEngine.Input.GetMouseButton(2) ||
-                EditorApplication.isCompiling ||
-                EditorApplication.isUpdating ||
-                EditorApplication.isPlayingOrWillChangePlaymode)
-            {
-                _lastActivity = EditorApplication.timeSinceStartup;
-            }
-        }
-        public static double TimeSinceLastActivity => EditorApplication.timeSinceStartup - _lastActivity;
-    }
-
     // Zip snapshot creation policy
     public enum ZipPolicy
     {
@@ -289,144 +122,7 @@ namespace AvatarSmartBackup
         public List<ManifestEntry> entries = new List<ManifestEntry>();
     }
 
-    internal static class FileUtilEx
-    {
-        public static string ProjectRoot => Directory.GetParent(Application.dataPath).FullName;
-        public static string ProjectName => new DirectoryInfo(ProjectRoot).Name;
-        public static string BackupRoot => Path.Combine(ProjectRoot, $"{Sanitize(ProjectName)} Backups");
-        public static string Sanitize(string s) { foreach (var c in Path.GetInvalidFileNameChars()) s = s.Replace(c, '_'); return s.Trim(); }
-        public static string MakeRelToProject(string abs) { var r = ProjectRoot.Replace('\\', '/'); var p = abs.Replace('\\', '/'); return p.StartsWith(r + "/") ? p.Substring(r.Length + 1) : p; }
-        public static string AssetToAbs(string ap) => Path.Combine(ProjectRoot, ap);
-
-        public static string MD5Of(string file)
-        {
-            using var md5 = MD5.Create();
-            using var stream = File.OpenRead(file);
-            var hash = md5.ComputeHash(stream);
-            var sb = new StringBuilder(hash.Length * 2);
-            foreach (var b in hash) sb.Append(b.ToString("x2"));
-            return sb.ToString();
-        }
-
-        public static string TryReadGuidFromMeta(string assetAbs)
-        {
-            try
-            {
-                string meta = assetAbs + ".meta";
-                if (!File.Exists(meta)) return null;
-                foreach (var line in File.ReadLines(meta))
-                {
-                    if (line.StartsWith("guid:", StringComparison.OrdinalIgnoreCase))
-                    {
-                        var val = line.Substring(5).Trim().Trim(':').Trim();
-                        return val;
-                    }
-                }
-            }
-            catch { }
-            return null;
-        }
-
-        public static void AtomicReplace(string tmp, string finalPath)
-        {
-            Directory.CreateDirectory(Path.GetDirectoryName(finalPath));
-            if (File.Exists(finalPath))
-            {
-                try { File.Replace(tmp, finalPath, null, true); }
-                catch { File.Delete(finalPath); File.Move(tmp, finalPath); }
-            }
-            else File.Move(tmp, finalPath);
-        }
-    }
-
     internal interface IBackupCollector { IEnumerable<string> CollectAbsolutePaths(BackupSettings s); }
-
-    internal static class CollectHelpers
-    {
-        static bool MatchesExt(string assetPath, string pat)
-        {
-            if (string.IsNullOrEmpty(pat)) return false;
-            pat = pat.Trim();
-            if (pat.StartsWith("*")) pat = pat.Substring(1);
-            if (!pat.StartsWith(".")) return false;
-            return assetPath.EndsWith(pat, StringComparison.OrdinalIgnoreCase);
-        }
-        public static bool PassesFolderFilters(string assetPath, BackupSettings s)
-        {
-            if (!assetPath.StartsWith("Assets/", StringComparison.OrdinalIgnoreCase)) return false;
-            if (s.includeFolders != null && s.includeFolders.Count > 0)
-            {
-                bool any = s.includeFolders.Any(f =>
-                {
-                    var t = (f ?? string.Empty).Trim();
-                    if (string.IsNullOrEmpty(t)) return false;
-                    if (t.StartsWith("Assets/", StringComparison.OrdinalIgnoreCase))
-                        return assetPath.StartsWith(t, StringComparison.OrdinalIgnoreCase);
-                    // Treat entries like ".anim" or "*.anim" as extension filters
-                    return MatchesExt(assetPath, t);
-                });
-                if (!any) return false;
-            }
-            if (s.excludeFolders != null && s.excludeFolders.Any(f =>
-            {
-                var t = (f ?? string.Empty).Trim();
-                if (string.IsNullOrEmpty(t)) return false;
-                if (t.StartsWith("Assets/", StringComparison.OrdinalIgnoreCase))
-                    return assetPath.StartsWith(t, StringComparison.OrdinalIgnoreCase);
-                return MatchesExt(assetPath, t);
-            })) return false;
-            return true;
-        }
-    }
-
-    // Tracks file changes incrementally using FileSystemWatcher to avoid full scans
-    internal static class IncrementalCollector
-    {
-        static readonly HashSet<string> Changed = new HashSet<string>(StringComparer.OrdinalIgnoreCase);
-        static FileSystemWatcher _watcher;
-
-        static IncrementalCollector()
-        {
-            try
-            {
-                string root = Path.Combine(FileUtilEx.ProjectRoot, "Assets");
-                if (Directory.Exists(root))
-                {
-                    _watcher = new FileSystemWatcher(root)
-                    {
-                        IncludeSubdirectories = true,
-                        NotifyFilter = NotifyFilters.FileName | NotifyFilters.DirectoryName | NotifyFilters.LastWrite
-                    };
-                    _watcher.Changed += OnChange;
-                    _watcher.Created += OnChange;
-                    _watcher.Deleted += OnChange;
-                    _watcher.Renamed += (s, e) =>
-                    {
-                        OnChange(s, new FileSystemEventArgs(WatcherChangeTypes.Changed, Path.GetDirectoryName(e.FullPath), e.Name));
-                        OnChange(s, new FileSystemEventArgs(WatcherChangeTypes.Changed, Path.GetDirectoryName(e.OldFullPath), e.OldName));
-                    };
-                    _watcher.EnableRaisingEvents = true;
-                }
-            }
-            catch { _watcher = null; }
-        }
-
-        static void OnChange(object sender, FileSystemEventArgs e)
-        {
-            lock (Changed) Changed.Add(e.FullPath);
-        }
-
-        public static IEnumerable<string> ConsumeChanges()
-        {
-            lock (Changed)
-            {
-                var arr = Changed.ToArray();
-                Changed.Clear();
-                return arr;
-            }
-        }
-    }
-
     // === Collectors (come prima) ===
     internal class VRCAssetsCollector : IBackupCollector
     {
@@ -574,43 +270,6 @@ namespace AvatarSmartBackup
                         string rel = FileUtilEx.MakeRelToProject(abs).Replace("\\", "/");
                         if (!CollectHelpers.PassesFolderFilters(rel, s)) continue;
                         yield return abs;
-                    }
-                }
-            }
-        }
-    }
-
-    internal static class IOThrottle
-    {
-        // Copia stream -> stream con throttle (MB/s). 0 = illimitato.
-        public static void CopyStreamThrottled(Stream src, Stream dst, int bufferBytes, int maxMBps, CancellationToken ct)
-        {
-            byte[] buffer = new byte[bufferBytes];
-            var sw = Stopwatch.StartNew();
-            long total = 0;
-            double maxBps = maxMBps <= 0 ? double.PositiveInfinity : maxMBps * 1024.0 * 1024.0;
-
-            while (true)
-            {
-                ct.ThrowIfCancellationRequested();
-                int r = src.Read(buffer, 0, buffer.Length);
-                if (r <= 0) break;
-                dst.Write(buffer, 0, r);
-                total += r;
-
-                if (maxMBps > 0)
-                {
-                    double elapsed = sw.Elapsed.TotalSeconds;
-                    if (elapsed > 0)
-                    {
-                        double bps = total / elapsed;
-                        if (bps > maxBps)
-                        {
-                            // tempo desiderato per scrivere 'total' a maxBps
-                            double desired = total / maxBps;
-                            int sleepMs = (int)Math.Max(1, (desired - elapsed) * 1000.0);
-                            Thread.Sleep(sleepMs);
-                        }
                     }
                 }
             }
